Part 4: Alternate JSON input using GET
Using Javascript/jQuery, I added a GET handler that accepts a variable
called “json” that allows a
user to point to an alternate JSON file. If the alternate JSON file is
valid, use that file as the
gallery’s input.  If the file is not valid it loads images.json